Fixed vercel update
