+++ apps/searchneu/app/catalog/[term]/[course]/page.tsx
@@ -4,7 +4,8 @@ import { getSectionsByCourseId } from "@/lib/dal/sections";
 import { getTrackedSectionIds } from "@/lib/dal/trackers";
 import { ExpandableDescription } from "@/components/catalog/ExpandableDescription";
 import { Separator } from "@/components/ui/separator";
-import { ExternalLink, Globe, GlobeLock } from "lucide-react";
+import { ExternalLink } from "lucide-react";
+import { TermLastUpdated } from "@/components/catalog/TermLastUpdated";
 import { Suspense } from "react";
 import { unstable_cache } from "next/cache";
 import { type Requisite } from "@sneu/scraper/types";
@@ -91,29 +92,10 @@ export default async function Page(props: {
           >
             {formatCreditRangeString(course.minCredits, course.maxCredits)}
           </h2>
-          <span className="text-neu6 flex max-w-20 items-center gap-1 sm:max-w-full">
-            {isTermActive ? (
-              <>
-                <Globe className="size-4" />
-                <h2
-                  style={{ lineHeight: 1.3 }}
-                  className="text-neu6 text-xs italic md:text-sm"
-                >
-                  {formatLastUpdatedString(term?.updatedAt)}
-                </h2>
-              </>
-            ) : (
-              <>
-                <GlobeLock className="size-4" />
-                <h2
-                  style={{ lineHeight: 1.3 }}
-                  className="text-neu6 text-xs italic md:text-sm"
-                >
-                  {"Last updated " + term.updatedAt.toLocaleDateString()}
-                </h2>
-              </>
-            )}
-          </span>
+          <TermLastUpdated
+            updatedAt={term.updatedAt}
+            isTermActive={isTermActive}
+          />
         </div>
       </div>
 
@@ -234,37 +216,6 @@ function formatCreditRangeString(minCredits: string, maxCredits: string) {
   }
 
   return creditRange + " " + creditLabel;
-}
-
-function formatLastUpdatedString(date: Date) {
-  const seconds = Math.floor((new Date().getTime() - date.getTime()) / 1000);
-  let str = "Updated ";
-
-  if (seconds < 0) {
-    str += "in the future???";
-    return str;
-  }
-
-  if (seconds < 60) {
-    str += "less than a minute ago";
-    return str;
-  }
-
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) {
-    str += minutes === 1 ? "1 minute ago" : `${minutes} minutes ago`;
-    return str;
-  }
-
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) {
-    str += hours === 1 ? "1 hour ago" : `${hours} hours ago`;
-    return str;
-  }
-
-  const days = Math.floor(hours / 24);
-  str += days === 1 ? "1 day ago" : `${days} days ago`;
-  return str;
 }
 
 function SectionsTableSkeleton() {
